Prefixes installer config fields based on backends
Some fields in the config are only used by WiX atm. This makes
the non shared fields optionnal.

A side effect of this change is that I had to move the logic handling
the image files from the opam frontend to the WiX backend, removing them
from the bundle directory in the process as this was not necessary.

Signed-off-by: Nathan Rebours <nathan.rebours@ocamlpro.com>