Support installing multiple binaries in makeself backend
This changes the configuration field to allow setting multiple
binaries and changes the opam frontend to bundle all binaries.

I temporarily added an error case in the WiX backend to exit
when more or less than one binary are in the config. This should soon
be fixed by @ddeclerck updates on the WiX backend.

Signed-off-by: Nathan Rebours <nathan.rebours@ocamlpro.com>