Allow plugins libs to overlap within a bundle
This allows to define multiple plugins to point to the same library
folder. This is something that's possible to do since two different
plugins can point to two different libs in the same package but those
are both in the same directory in the `lib/` folder. Since we install
folders rather than fine grained libraries when installing plugins
(because we in fact install symlink to those directories), we need to
support this somehow.

This makes it so libraries are installed only if there is no library
sitting there already.
This also allow us to let plugins from different installers share
dependencies.
The two cases are handled slightly differently:
1. the plugin's library must not be present before the install but
   during the installation, it's only installed once.
2. dependency libraries can be present before installation, we simply
  check they are a directory or a symlink and do not overwrite them if
  that's the case. This could potentially be later refined with checks
  on the lib META files to ensure the version is consistent for
  instance.